Fix signed/unsigned mismatch in zGothicAPI.cpp to make it compile under /W4 /WX flags

See merge request union-framework/gothic-api!29

+++ ZenGin/zGothicAPI.cpp
@@ -56,7 +56,7 @@ int GetGameVersion() {
 
 
 static const char* GothicNamespace = nullptr;
-static uint32_t GothicNamespaceLength = 0;
+static size_t GothicNamespaceLength = 0;
 
 
 inline void DisableNamespace( char* where ) {
